new_installer: Add abstract base classes for Tee, ChildInfo
