+++ lib/spack/spack/new_installer_posix.py
@@ -2,7 +2,7 @@
 #
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
 
-"""POSIX-specific terminal state and stdin reader for the new_installer TUI."""
+"""POSIX-specific terminal state, stdin reader, IPC channels, and job scheduling."""
 
 import fcntl
 import io
@@ -13,25 +13,21 @@ import signal
 import sys
 import tempfile
 import termios
-import threading
 import tty
 import warnings
-from multiprocessing import Process
 from multiprocessing.connection import Connection
 from typing import TYPE_CHECKING, Any, Callable, Optional, Tuple, Union
 
-import spack.database
 import spack.llnl.util.tty
 import spack.spec
-import spack.util.lock
 from spack.llnl.util.tty.log import _is_background_tty, ignore_signal
 from spack.new_installer_base import (
     OUTPUT_BUFFER_SIZE,
     BaseTerminalState,
-    DatabaseAction,
-    FdInfo,
+    ChildInfo,
     JobServerBase,
     StdinReaderBase,
+    Tee,
 )
 
 if TYPE_CHECKING:
@@ -198,86 +194,11 @@ class PosixTerminalState(BaseTerminalState):
         return not _is_background_tty(sys.stdin)
 
 
-class PosixChildInfo(DatabaseAction):
-    """Information about a child process."""
-
-    __slots__ = ("proc", "output_r_conn", "state_r_conn", "control_w_conn", "explicit", "log_path")
-
-    def __init__(
-        self,
-        proc: Process,
-        spec: spack.spec.Spec,
-        output_r_conn: Connection,
-        state_r_conn: Connection,
-        control_w_conn: Connection,
-        log_path: str,
-        explicit: bool = False,
-    ) -> None:
-        self.proc = proc
-        self.spec = spec
-        self.output_r_conn = output_r_conn
-        self.state_r_conn = state_r_conn
-        self.control_w_conn = control_w_conn
-        self.log_path = log_path
-        self.explicit = explicit
-        self.prefix_lock: Optional[spack.util.lock.Lock] = None
-
-    def save_to_db(self, db: spack.database.Database) -> None:
-        return db._add(self.spec, explicit=self.explicit)
-
-    def register_with_selector(self, selector: selectors.BaseSelector, pid: int) -> None:
-        """Register output, state, and sentinel channels with the selector."""
-        selector.register(self.output_r_conn.fileno(), selectors.EVENT_READ, FdInfo(pid, "output"))
-        selector.register(self.state_r_conn.fileno(), selectors.EVENT_READ, FdInfo(pid, "state"))
-        selector.register(self.proc.sentinel, selectors.EVENT_READ, FdInfo(pid, "sentinel"))
-
-    def close(self, selector: selectors.BaseSelector) -> int:
-        """Unregister and close file descriptors, and join the child process.
-        Returns the exit code of the child process."""
-        try:
-            selector.unregister(self.output_r_conn.fileno())
-        except KeyError:
-            pass
-        try:
-            selector.unregister(self.state_r_conn.fileno())
-        except KeyError:
-            pass
-        try:
-            selector.unregister(self.proc.sentinel)
-        except (KeyError, ValueError):
-            pass
-        self.output_r_conn.close()
-        self.state_r_conn.close()
-        self.control_w_conn.close()
-        self.proc.join()
-        exit_code = self.proc.exitcode
-        assert exit_code is not None, "Finished build should have exit code set"
-        if hasattr(self.proc, "close"):  # No known equivalent in Python 3.6
-            self.proc.close()
-        return exit_code
+class PosixChildInfo(ChildInfo):
+    """Posix Compatible class for Child Information"""
 
 
-class PosixTee:
-    """Emulates ./build 2>&1 | tee build.log. The output is sent both to a log file and the parent
-    process (if echoing is enabled). The control_fd is used to enable/disable echoing."""
-
-    def __init__(self, control: Connection, parent: Connection, log_path: str) -> None:
-        self.control = control
-        self.parent = parent
-        # sys.stdout and sys.stderr may have been replaced with file objects under pytest, so
-        # redirect their file descriptors in addition to the original fds 1 and 2.
-        fds = {sys.stdout.fileno(), sys.stderr.fileno(), 1, 2}
-        self.saved_fds = {fd: os.dup(fd) for fd in fds}
-        #: The path of the log file
-        self.log_path = log_path
-        log_file = open(self.log_path, "ab")
-        r, w = os.pipe()
-        self.tee_thread = threading.Thread(target=self.run, args=(r, log_file), daemon=True)
-        self.tee_thread.start()
-        for fd in fds:
-            os.dup2(w, fd)
-        os.close(w)
-
+class PosixTee(Tee):
     def run(self, log_r: int, log_file: io.BufferedWriter) -> None:
         """Forward log_r to log_file and parent (if echoing is enabled).
         Echoing is enabled and disabled by reading from control."""
@@ -312,21 +233,6 @@ class PosixTee:
             pass
         finally:
             os.close(log_r)
-
-    def close(self) -> None:
-        # Closing stdout and stderr should close the last reference to the write end of the pipe,
-        # causing the tee thread to wake up, flush the last data, and exit. We restore stdout and
-        # stderr, because between sys.exit and the actual process exit buffers may be flushed, and
-        # can cause exit code 120 (witnessed under pytest+coverage on macOS).
-        sys.stdout.flush()
-        sys.stderr.flush()
-        for fd, saved_fd in self.saved_fds.items():
-            os.dup2(saved_fd, fd)
-            os.close(saved_fd)
-        self.tee_thread.join()
-        # Only then close the other fds.
-        self.control.close()
-        self.parent.close()
 
 
 class PosixJobServer(JobServerBase):
